# API_Gateway_Authentication_Service/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subscription Management API",
        "version": "1.0.0-SNAPSHOT",
...
        "contact": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            }
        },
        "license": {
            "name": "MIT",
            "url": "https://spdx.org/licenses/MIT.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "notFoundResponseText": {
                "type": "string",
                "maxLength": 2000,
                "pattern": "^[ \\w\\.!?,;:\\-=><\\/&%\\$\\+\\*~#]*$",
                "example": "Not Found"
            },
...
            "attrAuthzPassword": {
                "type": "string",
                "example": "pr@ga!98",
                "maxLength": 30,
                "pattern": "^[\\w\\-!#\\$?\\+\\*\\.,;&\\/:@]{8,}$"
            },
            "loginForm": {
                "description": "the user form input",
                "type": "object",
                "additionalProperties": false,
                "required": ["username", "password"],
                "properties": {
                    "username": {
                        "$ref": "#/components/schemas/attributeUserName"
                    },
                    "password": {
                        "$ref": "#/components/schemas/attrAuthzPassword"
                    }
                }
            },
...
            "attributeUserName": {
                "type": "string",
                "maxLength": 50,
                "pattern": "^[- \\w\\.!?,;:-=><\\/()&%$+*~#]*$",
                "example": "Davide"
            },
...
            "newUserInput": {
                "description": "input for creating a new user",
                "type": "object",
                "additionalProperties": false,
                "required": ["email", "password", "name"],
                "properties": {
                    "email": {
                        "$ref": "#/components/schemas/attrUserEmail"
                    },
                    "password": {
                        "$ref": "#/components/schemas/attrAuthzPassword"
                    },
                    "name": {
                        "$ref": "#/components/schemas/attributeUserName"
                    }
                }
            },
...
            "patchUserInput": {
                "description": "input for creating a new user",
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "email": {
                        "$ref": "#/components/schemas/attrUserEmail"
                    },
                    "password": {
                        "$ref": "#/components/schemas/attrAuthzPassword"
                    },
                    "roles": {
                        "$ref": "#/components/schemas/attrRoles"
                    }
                }
            }
        },
        "responses": {
...
            "CreateUserItem": {
                "description": "Created",
                "headers": {
                    "Location": {
                        "schema": {
                            "type": "string",
                            "format": "uri"
                        },
                        "description": "the id (location) of the newly created user"
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/userData"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/userItem"
                        }
                    }
                }
            },
            "OkUserCollection": {
                "description": "the user collection",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/usersDataCollection"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/usersCollection"
                        }
                    }
                }
            },
            "OKUser": {
                "description": "the user item",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/userData"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/userItem"
                        }
                    }
                },
                "links": {
                    "self": {
                        "operationId": "getUser",
                        "parameters": {
                            "id": "$response.body#/id"
                        }
                    },
                    "login": {
                        "operationId": "login"
                    },
                    "collection": {
                        "operationId": "getUsers"
                    },
                    "update": {
                        "operationId": "updateUser",
                        "parameters": {
                            "id": "$response.body#/id"
                        }
                    },
                    "delete": {
                        "operationId": "deleteUser",
                        "parameters": {
                            "id": "$response.body#/id"
                        }
                    }
                }
            },
            "PreconditionFailedUserItem": {
                "description": "if-match header caused conflict",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/userData"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/userItem"
                        }
                    }
                }
            },
            "OK": {
                "description": "Ok.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/serverResponseText"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Not Found.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/notFoundResponseText"
                        }
                    }
                }
            },
...
            "BadRequest": {
                "description": "Bad Request.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/badRequestText"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized. You must first authenticate in order to perform this request.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/unauthorizedText"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden. You do not have enough permissions to perform this request.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/forbiddenText"
                        }
                    }
                }
            },
            "Conflict": {
                "description": "Conflict. The request could not be completed due to a conflict with the current state of the target resource.",
                "content": {
                    "text/plain": {
                        "schema": {
                            "$ref": "#/components/schemas/conflictText"
                        }
                    }
                }
            },
...
            "OkPlanCollection": {
                "description": "the plan collection",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/plansDataCollection"
                        }
                    },
                    "application/vnd.pagsousa.hyper+json": {
                        "schema": {
                            "$ref": "#/components/schemas/plansCollection"
                        }
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "summary": "authenticates a user",
                "description": "returns the authentication ID token (JWT) to use in further API calls.",
                "tags": ["Authz"],
                "security": [],
                "requestBody": {
                    "description": "login data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/loginForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "a JSON web token (JWT) required to login.",
                        "content": {
                            "application/jwt": {
                                "schema": {
                                    "type": "string",
                                    "maxLength": 1024,
                                    "pattern": "^[A-Za-z0-9-_=]+\\.[A-Za-z0-9-_=]+\\.?[A-Za-z0-9-_.+/=]*$",
                                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwczovL3NwYWNlYXBpLmNvbS9ncmFwaHFsIjp7InJvbGVzIjpbImFzdHJvbmF1dCJdLCJwZXJtaXNzaW9ucyI6WyJyZWFkOm93bl91c2VyIl19LCJpYXQiOjE1OTQyNTI2NjMsImV4cCI6MTU5NDMzOTA2Mywic3ViIjoiNjc4OTAifQ.Z1JPE53ca1JaxwDTlnofa3hwpS0PGdRLUMIrC7M3FCI"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "wrong Password"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "get all users",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "returns all users",
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "first",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "maximum": 500,
                            "format": "int32"
                        },
                        "description": "the first user to return",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "n",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 6,
                            "maximum": 500,
                            "format": "int32"
                        },
                        "description": "the number of users to return",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkUserCollection"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
        "/users/{userID}": {
            "get": {
                "description": "returns one user",
                "summary": "get one user",
                "tags": ["Users"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "schema": {
                            "$ref": "#/components/schemas/attributeUserName"
                        },
                        "description": "the user ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OKUser"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "description": "creates one user",
                "summary": "creates one user",
                "tags": ["Users"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "schema": {
                            "$ref": "#/components/schemas/attributeUserName"
                        },
                        "description": "the user ID",
                        "required": true
                    }
                ],
                "security": [],
                "requestBody": {
                    "description": "the user data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/newUserInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/CreateUserItem"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            },
            "delete": {
                "description": "deletes one user",
                "summary": "deletes one user",
                "tags": ["Users"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "schema": {
                            "$ref": "#/components/schemas/attributeUserName"
                        },
                        "description": "the user ID",
                        "required": true
                    },
                    {
                        "in": "header",
                        "name": "if-match",
                        "description": "Revision control",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/OK"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "412": {
                        "$ref": "#/components/responses/PreconditionFailedUserItem"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    }
                }
            },
            "patch": {
                "description": "updates one user",
                "summary": "updates one user",
                "tags": ["Users"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "schema": {
                            "$ref": "#/components/schemas/attributeUserName"
                        },
                        "description": "the user ID",
                        "required": true
                    },
                    {
                        "in": "header",
                        "name": "if-match",
                        "description": "Revision control",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "the user data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/patchUserInput"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },

                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "412": {
                        "$ref": "#/components/responses/PreconditionFailedUserItem"
                    },
                    "200": {
                        "$ref": "#/components/responses/OKUser"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    }
                }
            }
        },
...
        "/plans": {
            "get": {
                "description": "returns all plans",
                "summary": "returns all plans",
                "tags": ["Plans"],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "first",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "maximum": 500,
                            "format": "int32"
                        },
                        "description": "the first user to return",
                        "required": false
                    },
                    {
                        "in": "query",
                        "name": "n",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 6,
                            "maximum": 500,
                            "format": "int32"
                        },
                        "description": "the number of users to return",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/OkPlanCollection"
                    }
                }
            }
        },
...
    },
...
}